[fix] fix import

+++ src/main/java/com/bb/blog/entity/Post.java
@@ -1,5 +1,10 @@
 package com.bb.blog.entity;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
+import java.time.LocalDateTime;
 @Entity
 @Table(name = "posts")
 @Data
@@ -8,7 +13,7 @@ package com.bb.blog.entity;
 public class Post {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "user_id", nullable = false)
     private Integer userId;
